func: implemento la obtencion de la transaccion en el controller api

+++ src/main/java/org/example/repository/TransactionRepository.java
@@ -3,6 +3,6 @@ import org.example.model.Transaction;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import java.util.List;
 
-public interface TransactionRepository extends MongoRepository<Transaction,Integer> {
+public interface TransactionRepository extends MongoRepository<Transaction,Long> {
     List<Transaction> findByType(String type);
 }
